# openapi.yaml
openapi: 3.0.3
info:
  title: Kick Accounting API
  version: 1.0.0
  description: >
    Backend accounting platform with General Ledger, journal entries,
    and financial reports (Trial Balance, P&L, Balance Sheet).
servers:
  - url: /api
    description: API base path
paths:
...
  /journal-entries:
    get:
      tags: [Journal Entries]
      summary: List journal entries (paginated)
      operationId: listJournalEntries
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          description: Filter from date (YYYY-MM-DD)
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          description: Filter to date (YYYY-MM-DD)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        "200":
          description: Paginated journal entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/JournalEntryLine"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
              example:
                data:
                  - id: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                    date: "2025-01-02"
                    accountId: "a1b2c3d4-0001-4000-8000-000000000001"
                    accountCode: "10100"
                    accountName: "Bank Account - Chase Business Checking - 0205"
                    description: "Herman Miller Chair"
                    debitAmount: null
                    creditAmount: 1650.00
                    referenceNumber: "JE-2D3D28BD"
                    counterparty: null
                    createdAt: "2025-06-15T10:00:00.000Z"
                    updatedAt: "2025-06-15T10:00:00.000Z"
                meta:
                  total: 412
                  page: 1
                  limit: 50
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "startDate: Invalid"

    post:
      tags: [Journal Entries]
      summary: Create a journal entry
      operationId: createJournalEntry
      description: >
        Creates a balanced journal entry. Validated against accounting spec invariants:
        (1) Each line has exactly one of debitAmount or creditAmount, never both.
        (2) Sum of debits must equal sum of credits (double-entry balance).
        (3) All referenced accountIds must exist.
        (4) No duplicate accountIds within the same entry.
        (5) Amounts must be positive (> 0).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJournalEntry"
            examples:
              simple-purchase:
                summary: Simple 2-line purchase
                value:
                  date: "2025-03-10"
                  description: "Monthly web hosting"
                  counterparty: "Hetzner"
                  lines:
                    - accountId: "a1b2c3d4-0001-4000-8000-000000000001"
                      debitAmount: 49.99
                    - accountId: "a1b2c3d4-0001-4000-8000-000000000002"
                      creditAmount: 49.99
              payroll-with-taxes:
                summary: Payroll with employer tax — 3 lines
                value:
                  date: "2025-03-31"
                  description: "March 2025 payroll"
                  counterparty: "Gusto"
                  lines:
                    - accountId: "a1b2c3d4-0001-4000-8000-000000000010"
                      debitAmount: 8500.00
                    - accountId: "a1b2c3d4-0001-4000-8000-000000000011"
                      debitAmount: 650.25
                    - accountId: "a1b2c3d4-0001-4000-8000-000000000012"
                      creditAmount: 9150.25
      responses:
        "201":
          description: Journal entry created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/JournalEntryLine"
              examples:
                simple-purchase:
                  summary: Response for simple 2-line purchase
                  value:
                    data:
                      - id: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                        date: "2025-03-10"
                        accountId: "a1b2c3d4-0001-4000-8000-000000000001"
                        accountCode: "62100"
                        accountName: "Web Hosting"
                        description: "Monthly web hosting"
                        debitAmount: 49.99
                        creditAmount: null
                        referenceNumber: "JE-A7F2B91D"
                        counterparty: "Hetzner"
                        createdAt: "2025-03-10T14:30:00.000Z"
                        updatedAt: "2025-03-10T14:30:00.000Z"
                      - id: "f47ac10b-58cc-4372-a567-0e02b2c3d480"
                        date: "2025-03-10"
                        accountId: "a1b2c3d4-0001-4000-8000-000000000002"
                        accountCode: "10100"
                        accountName: "Bank Account - Chase Business Checking - 0205"
                        description: "Monthly web hosting"
                        debitAmount: null
                        creditAmount: 49.99
                        referenceNumber: "JE-A7F2B91D"
                        counterparty: "Hetzner"
                        createdAt: "2025-03-10T14:30:00.000Z"
                        updatedAt: "2025-03-10T14:30:00.000Z"
        "400":
          description: Validation error (invalid input, duplicate accounts, missing accounts)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "lines: Each line must have exactly one of debitAmount or creditAmount"
        "422":
          description: Double-entry violation (debits != credits)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "lines: Double-entry violation: total debits (500.00) must equal total credits (499.00)"
  /journal-entries/{ref}:
    parameters:
      - name: ref
        in: path
        required: true
        schema:
          type: string
        description: Reference number (e.g. JE-2D3D28BD)

    get:
      tags: [Journal Entries]
      summary: Get journal entry by reference number
      operationId: getJournalEntry
      responses:
        "200":
          description: Journal entry lines
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/JournalEntryLine"
              example:
                data:
                  - id: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                    date: "2025-03-10"
                    accountId: "a1b2c3d4-0001-4000-8000-000000000001"
                    accountCode: "62100"
                    accountName: "Web Hosting"
                    description: "Monthly web hosting"
                    debitAmount: 49.99
                    creditAmount: null
                    referenceNumber: "JE-A7F2B91D"
                    counterparty: "Hetzner"
                    createdAt: "2025-03-10T14:30:00.000Z"
                    updatedAt: "2025-03-10T14:30:00.000Z"
                  - id: "f47ac10b-58cc-4372-a567-0e02b2c3d480"
                    date: "2025-03-10"
                    accountId: "a1b2c3d4-0001-4000-8000-000000000002"
                    accountCode: "10100"
                    accountName: "Bank Account - Chase Business Checking - 0205"
                    description: "Monthly web hosting"
                    debitAmount: null
                    creditAmount: 49.99
                    referenceNumber: "JE-A7F2B91D"
                    counterparty: "Hetzner"
                    createdAt: "2025-03-10T14:30:00.000Z"
                    updatedAt: "2025-03-10T14:30:00.000Z"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Journal entry not found"

    put:
      tags: [Journal Entries]
      summary: Update a journal entry
      operationId: updateJournalEntry
      description: >
        Replaces all lines of an existing journal entry.
        Same validation rules as create: double-entry balance, no duplicate accounts, all accounts must exist.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJournalEntry"
      responses:
        "200":
          description: Updated journal entry
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/JournalEntryLine"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Journal entry JE-XXXXXXXX not found"

    delete:
      tags: [Journal Entries]
      summary: Delete a journal entry
      operationId: deleteJournalEntry
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Journal entry JE-XXXXXXXX not found"
...
  /export/download:
    get:
      tags: [Export]
      summary: Export GL to Excel
      operationId: exportExcel
      description: >
        Downloads the General Ledger for the given date range as an .xlsx file,
        grouped by account with Beginning/Ending Balance rows, running balance,
        and formatted amounts.
        Columns: Date, Source, Reference, Description, Counterparty, Split, Amount, Balance.
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Period start (YYYY-MM-DD)
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Period end (YYYY-MM-DD)
      responses:
        "200":
          description: Excel file download
          headers:
            Content-Disposition:
              schema:
                type: string
                example: 'attachment; filename="general_ledger_2025-01-01_2025-12-31.xlsx"'
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /export/google:
    post:
      tags: [Export]
      summary: Export GL to Google Sheets
      operationId: exportGoogleSheets
      description: >
        Creates a new Google Spreadsheet with the General Ledger for the given
        date range, using the same layout as the Excel export: grouped by account
        with Beginning/Ending Balance rows, running balance, and formatted amounts.
        Columns: Date, Source, Reference, Description, Counterparty, Split, Amount, Balance.
        Optionally shares it with a specified email address.
        Requires Google service account credentials (GOOGLE_SA_KEY_PATH or GOOGLE_SA_KEY_JSON).
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Period start (YYYY-MM-DD)
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Period end (YYYY-MM-DD)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleExportOptions"
            example:
              shareWith: "pawel@example.com"
      responses:
        "200":
          description: Created spreadsheet info
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/GoogleExportResult"
              example:
                data:
                  spreadsheetId: "1BxiMVs0XRA5nFMdKvBdBZjgmUUqptlbs74OgVE2upms"
                  spreadsheetUrl: "https://docs.google.com/spreadsheets/d/1BxiMVs0XRA5nFMdKvBdBZjgmUUqptlbs74OgVE2upms/edit"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Google API error (credentials not configured, quota exceeded, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Google credentials not configured. Set GOOGLE_SA_KEY_PATH or GOOGLE_SA_KEY_JSON."
...
components:
  schemas:
...
    JournalEntryLine:
      type: object
      required: [id, date, accountId, referenceNumber, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        accountId:
          type: string
          format: uuid
        accountCode:
          type: string
          example: "10100"
        accountName:
          type: string
          example: "Bank Account - Chase Business Checking - 0205"
        description:
          type: string
          nullable: true
        debitAmount:
          type: number
          nullable: true
          description: Decimal amount (e.g. 500.00), null if this is a credit line
        creditAmount:
          type: number
          nullable: true
          description: Decimal amount (e.g. 500.00), null if this is a debit line
        referenceNumber:
          type: string
          example: "JE-2D3D28BD"
        counterparty:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    JournalEntryLineInput:
      type: object
      required: [accountId]
      properties:
        accountId:
          type: string
          format: uuid
        debitAmount:
          type: number
          minimum: 0.01
          description: Provide debitAmount OR creditAmount, not both
        creditAmount:
          type: number
          minimum: 0.01
    CreateJournalEntry:
      type: object
      required: [date, lines]
      properties:
        date:
          type: string
          format: date
          example: "2025-01-15"
        description:
          type: string
          maxLength: 500
          example: "Office supplies purchase"
        counterparty:
          type: string
          maxLength: 255
          example: "Staples"
        lines:
          type: array
          minItems: 2
          items:
            $ref: "#/components/schemas/JournalEntryLineInput"
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 412
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
...
    GoogleExportOptions:
      type: object
      properties:
        shareWith:
          type: string
          format: email
          description: Email address to share the created spreadsheet with (writer access)
    GoogleExportResult:
      type: object
      properties:
        spreadsheetId:
          type: string
        spreadsheetUrl:
          type: string
          format: uri
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message